Add files via upload
First check in. The class CreateDSM2ChanPolygons successfully automates the creation of polygons based on CSDP centerlines that can be used to estimate channel volume using GIS. Polygons may need to be enlarged in the transverse direction to make sure the boundaries are touching the levees or are inside the levees at a given elevation. Output is written to a CSDP network file, which can be exported to WKT. Results for some channels are way too big, due to inconsistencies in the creation of the centerlines. These polygons will have to be edited using a new feature in the CSDP.